+++ tests/test_action_output.py
@@ -13,7 +13,32 @@ class OutputTestCase(TerraformBaseActionTestCase):
 
     @mock.patch("output.os.chdir")
     @mock.patch("lib.action.Terraform.output")
-    def test_run(self, mock_output, mock_chdir):
+    def test_run_state_file(self, mock_output, mock_chdir):
+        action = self.get_action_instance({})
+        # Declare test input values
+        test_plan_path = "/terraform"
+        test_state_file_path = "/path/to/state/file"
+        test_terraform_exec = "/usr/bin/terraform"
+
+        # Declare test Terraform.output return values
+        expected_result = "result"
+        mock_output.return_value = expected_result
+
+        mock_chdir.return_value = "success"
+
+        # Execute the run function
+        result = action.run(test_plan_path, test_state_file_path, test_terraform_exec)
+
+        # Verify the results
+        self.assertEqual(result, expected_result)
+        self.assertEqual(action.terraform.terraform_bin_path, test_terraform_exec)
+        self.assertTrue(mock_output.called)
+        mock_chdir.assert_called_with(test_plan_path)
+        mock_output.assert_called_with(state=test_state_file_path)
+
+    @mock.patch("output.os.chdir")
+    @mock.patch("lib.action.Terraform.output")
+    def test_run_no_state_file(self, mock_output, mock_chdir):
         action = self.get_action_instance({})
         # Declare test input values
         test_plan_path = "/terraform"
@@ -26,10 +51,11 @@ class OutputTestCase(TerraformBaseActionTestCase):
         mock_chdir.return_value = "success"
 
         # Execute the run function
-        result = action.run(test_plan_path, test_terraform_exec)
+        result = action.run(test_plan_path, None, test_terraform_exec)
 
         # Verify the results
         self.assertEqual(result, expected_result)
         self.assertEqual(action.terraform.terraform_bin_path, test_terraform_exec)
         self.assertTrue(mock_output.called)
         mock_chdir.assert_called_with(test_plan_path)
+        mock_output.assert_called_with(state=None)